obnubilate: output the means for each systematic

+++ MainAnalysis/src/obnubilate.C
@@ -288,7 +288,7 @@ int obnubilate(char const* config, char const* selections, char const* output) {
 
         total->apply(sqrt_);
 
-        for (int64_t i = 0; i <= sets.size(); ++i) {
+        for (size_t i = 0; i <= sets.size(); ++i) {
             for (int64_t j = 0; j < total->size(); ++j) {
                 double avg;
